demo

+++ remind-parent/remind-guava/src/main/java/com/fsmeeting/guava/cache/IUserService.java
@@ -9,5 +9,5 @@ import com.fsmeeting.model.User;
  */
 public interface IUserService {
 
-    User getById(Integer userId);
+    User getById(Long userId);
 }
